ADD: 커뮤니티 게시글 좋아요 Api

+++ src/main/java/lostark/todo/domainV2/board/community/dto/CommunitySearchResponse.java
@@ -20,7 +20,7 @@ public class CommunitySearchResponse {
 
     private boolean myPost;
 
-    private int likeCount;
+    private long likeCount;
 
     private boolean myLike;
